No need for the whole package - move pack management related utility functions in st2common.util.pack_management module.
