+++ contrib/packs/tests/test_action_download.py
@@ -31,8 +31,9 @@ from gitdb.exc import BadName
 from st2common.services import packs as pack_service
 from st2tests.base import BaseActionTestCase
 
-import st2common.pack_management.download
-from st2common.pack_management.download import eval_repo_url
+import st2common.util.pack_management
+from st2common.util.pack_management import eval_repo_url
+
 from pack_mgmt.download import DownloadGitRepoAction
 
 PACK_INDEX = {
@@ -219,7 +220,7 @@ class DownloadGitRepoActionTestCase(BaseActionTestCase):
 
         self.assertEqual(result, {'test': 'Success.'})
 
-    @mock.patch.object(st2common.pack_management.download, 'get_valid_versions_for_repo',
+    @mock.patch.object(st2common.util.pack_management, 'get_valid_versions_for_repo',
                        mock.Mock(return_value=['1.0.0', '2.0.0']))
     def test_run_pack_download_invalid_version(self):
         self.repo_instance.commit.side_effect = lambda ref: None
@@ -235,38 +236,38 @@ class DownloadGitRepoActionTestCase(BaseActionTestCase):
         action = self.get_action_instance()
 
         # Version is satisfied
-        st2common.pack_management.download.CURRENT_STACKSTROM_VERSION = '2.0.0'
+        st2common.util.pack_management.CURRENT_STACKSTROM_VERSION = '2.0.0'
 
         result = action.run(packs=['test3'], abs_repo_base=self.repo_base)
         self.assertEqual(result['test3'], 'Success.')
 
         # Pack requires a version which is not satisfied by current StackStorm version
-        st2common.pack_management.download.CURRENT_STACKSTROM_VERSION = '2.2.0'
+        st2common.util.pack_management.CURRENT_STACKSTROM_VERSION = '2.2.0'
         expected_msg = ('Pack "test3" requires StackStorm ">=1.6.0, <2.2.0", but '
                         'current version is "2.2.0"')
         self.assertRaisesRegexp(ValueError, expected_msg, action.run, packs=['test3'],
                                 abs_repo_base=self.repo_base)
 
-        st2common.pack_management.download.CURRENT_STACKSTROM_VERSION = '2.3.0'
+        st2common.util.pack_management.CURRENT_STACKSTROM_VERSION = '2.3.0'
         expected_msg = ('Pack "test3" requires StackStorm ">=1.6.0, <2.2.0", but '
                         'current version is "2.3.0"')
         self.assertRaisesRegexp(ValueError, expected_msg, action.run, packs=['test3'],
                                 abs_repo_base=self.repo_base)
 
-        st2common.pack_management.download.CURRENT_STACKSTROM_VERSION = '1.5.9'
+        st2common.util.pack_management.CURRENT_STACKSTROM_VERSION = '1.5.9'
         expected_msg = ('Pack "test3" requires StackStorm ">=1.6.0, <2.2.0", but '
                         'current version is "1.5.9"')
         self.assertRaisesRegexp(ValueError, expected_msg, action.run, packs=['test3'],
                                 abs_repo_base=self.repo_base)
 
-        st2common.pack_management.download.CURRENT_STACKSTROM_VERSION = '1.5.0'
+        st2common.util.pack_management.CURRENT_STACKSTROM_VERSION = '1.5.0'
         expected_msg = ('Pack "test3" requires StackStorm ">=1.6.0, <2.2.0", but '
                         'current version is "1.5.0"')
         self.assertRaisesRegexp(ValueError, expected_msg, action.run, packs=['test3'],
                                 abs_repo_base=self.repo_base)
 
         # Version is not met, but force=true parameter is provided
-        st2common.pack_management.download.CURRENT_STACKSTROM_VERSION = '1.5.0'
+        st2common.util.pack_management.CURRENT_STACKSTROM_VERSION = '1.5.0'
         result = action.run(packs=['test3'], abs_repo_base=self.repo_base, force=True)
         self.assertEqual(result['test3'], 'Success.')
